requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.3
cron-descriptor==1.4.3
Django==4.2.9
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psycopg2
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
